v0.2.3 external commit

+++ R/initialise_project.R
@@ -23,7 +23,12 @@ initialise_project <- function(path,
             overwrite = FALSE, recursive = TRUE,
             copy.mode = TRUE)
 
-  file.remove(paste0(path,"/.git/config"))
+  if(file.exists(paste0(path,"/.git/config")))
+  {
+
+    file.remove(paste0(path,"/.git/config"))
+
+  }
 
   file.rename(from=paste0(path,"/new_project_template.Rproj"),
               to=paste0(path,"/",
